docs/requirements.txt
Sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
sphinxcontrib-bibtex>=2.4.2
sphinxcontrib-katex>=0.8.6